Files and folder refactor
