+++ WIP/axonsegpy/GUI.py
