+++ WIP/user/input/algo.json
