--- axonsegpy/cythonLib/__init__.py
@@ -1,2 +1,0 @@
-from cythonLib import minima
-from cythonLib import GenMask
