--- axonsegpy/lib/__init__.py
@@ -1,2 +1,0 @@
-from lib import minima
-from lib import GenMask
