+++ axonsegpy/lib/compiledLib/__init__.py
@@ -1,0 +1,2 @@
+from lib.compiledLib import minima
+from lib.compiledLib import GenMask
