+++ axonsegpy/lib/cythonLib/__init__.py
@@ -1,0 +1,2 @@
+from lib.cythonLib import minima
+from lib.cythonLib import GenMask
